+++ common/buildcraft/core/blueprints/BuildingSlot.java
@@ -65,6 +65,6 @@ public abstract class BuildingSlot {
 
 	public abstract void readFromNBT(NBTTagCompound nbt, MappingRegistry registry) throws MappingNotFoundException;
 
-	public abstract double getEnergyRequirement();
+	public abstract int getEnergyRequirement();
 
 }
