+++ common/buildcraft/core/blueprints/BuildingSlotBlock.java
@@ -172,7 +172,7 @@ public class BuildingSlotBlock extends BuildingSlot {
 	}
 
 	@Override
-	public double getEnergyRequirement() {
+	public int getEnergyRequirement() {
 		return schematic.getEnergyRequirement(stackConsumed);
 	}
